feat: move k8sClusterReceiver to Agent
