+++ internal/otelcollector/config/metric/agent/receivers.go
@@ -2,6 +2,7 @@ package agent
 
 import (
 	"fmt"
+	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/metric"
 	"path/filepath"
 	"time"
 
@@ -21,6 +22,7 @@ func makeReceiversConfig(inputs inputSources, opts BuildOptions) Receivers {
 
 	if inputs.runtime {
 		receiversConfig.KubeletStats = makeKubeletStatsConfig()
+		receiversConfig.SingletonK8sClusterReceiverCreator = makeSingletonK8sClusterReceiverCreatorConfig(opts.AgentNamespace)
 	}
 
 	if inputs.istio {
@@ -51,6 +53,36 @@ func makeKubeletStatsConfig() *KubeletStatsReceiver {
 			K8sNodeMemoryPageFaults:      MetricConfig{Enabled: false},
 			K8sNodeMemoryRSS:             MetricConfig{Enabled: false},
 			K8sNodeMemoryWorkingSet:      MetricConfig{Enabled: false},
+		},
+	}
+}
+
+func makeSingletonK8sClusterReceiverCreatorConfig(gatewayNamespace string) *SingletonK8sClusterReceiverCreator {
+	metricsToDrop := K8sClusterMetricsConfig{
+		K8sContainerStorageRequest:          MetricConfig{false},
+		K8sContainerStorageLimit:            MetricConfig{false},
+		K8sContainerEphemeralStorageRequest: MetricConfig{false},
+		K8sContainerEphemeralStorageLimit:   MetricConfig{false},
+		K8sContainerRestarts:                MetricConfig{false},
+		K8sContainerReady:                   MetricConfig{false},
+		K8sNamespacePhase:                   MetricConfig{false},
+		K8sReplicationControllerAvailable:   MetricConfig{false},
+		K8sReplicationControllerDesired:     MetricConfig{false},
+	}
+
+	return &SingletonK8sClusterReceiverCreator{
+		AuthType: "serviceAccount",
+		LeaderElection: metric.LeaderElection{
+			LeaseName:      "telemetry-metric-gateway-k8scluster",
+			LeaseNamespace: gatewayNamespace,
+		},
+		SingletonK8sClusterReceiver: SingletonK8sClusterReceiver{
+			K8sClusterReceiver: K8sClusterReceiver{
+				AuthType:               "serviceAccount",
+				CollectionInterval:     "30s",
+				NodeConditionsToReport: []string{},
+				Metrics:                metricsToDrop,
+			},
 		},
 	}
 }
