+++ internal/otelcollector/config/metric/agent/receivers_test.go
@@ -29,9 +29,12 @@ func TestReceivers(t *testing.T) {
 	})
 
 	t.Run("runtime input enabled", func(t *testing.T) {
+		agentNamespace := "test-namespace"
 		collectorConfig := sut.Build([]telemetryv1alpha1.MetricPipeline{
 			testutils.NewMetricPipelineBuilder().WithRuntimeInput(true).Build(),
-		}, BuildOptions{})
+		}, BuildOptions{
+			AgentNamespace: agentNamespace,
+		})
 
 		require.Nil(t, collectorConfig.Receivers.PrometheusAppPods)
 		require.Nil(t, collectorConfig.Receivers.PrometheusIstio)
@@ -56,7 +59,32 @@ func TestReceivers(t *testing.T) {
 				K8sNodeMemoryWorkingSet:      MetricConfig{Enabled: false},
 			},
 		}
+
+		expectedMetricsToDrop := K8sClusterMetricsConfig{
+			K8sContainerStorageRequest:          MetricConfig{false},
+			K8sContainerStorageLimit:            MetricConfig{false},
+			K8sContainerEphemeralStorageRequest: MetricConfig{false},
+			K8sContainerEphemeralStorageLimit:   MetricConfig{false},
+			K8sContainerRestarts:                MetricConfig{false},
+			K8sContainerReady:                   MetricConfig{false},
+			K8sNamespacePhase:                   MetricConfig{false},
+			K8sReplicationControllerAvailable:   MetricConfig{false},
+			K8sReplicationControllerDesired:     MetricConfig{false},
+		}
+
 		require.Equal(t, expectedKubeletStatsReceiver, *collectorConfig.Receivers.KubeletStats)
+		singletonK8sClusterReceiverCreator := collectorConfig.Receivers.SingletonK8sClusterReceiverCreator
+		require.NotNil(t, singletonK8sClusterReceiverCreator)
+		require.Equal(t, "serviceAccount", singletonK8sClusterReceiverCreator.AuthType)
+		require.Equal(t, "telemetry-metric-gateway-k8scluster", singletonK8sClusterReceiverCreator.LeaderElection.LeaseName)
+		require.Equal(t, agentNamespace, singletonK8sClusterReceiverCreator.LeaderElection.LeaseNamespace)
+
+		k8sClusterReceiver := singletonK8sClusterReceiverCreator.SingletonK8sClusterReceiver.K8sClusterReceiver
+		require.Equal(t, "serviceAccount", k8sClusterReceiver.AuthType)
+		require.Equal(t, "30s", k8sClusterReceiver.CollectionInterval)
+		require.Len(t, k8sClusterReceiver.NodeConditionsToReport, 0)
+		require.Equal(t, expectedMetricsToDrop, k8sClusterReceiver.Metrics)
+		require.Equal(t, expectedMetricsToDrop, collectorConfig.Receivers.SingletonK8sClusterReceiverCreator.SingletonK8sClusterReceiver.K8sClusterReceiver.Metrics)
 	})
 
 	t.Run("prometheus input enabled", func(t *testing.T) {
