+++ internal/otelcollector/config/metric/config.go
@@ -6,3 +6,8 @@ type TransformProcessor struct {
 	ErrorMode        string                                `yaml:"error_mode"`
 	MetricStatements []config.TransformProcessorStatements `yaml:"metric_statements"`
 }
+
+type LeaderElection struct {
+	LeaseName      string `yaml:"lease_name"`
+	LeaseNamespace string `yaml:"lease_namespace"`
+}
