+++ internal/otelcollector/config/metric/gateway/config.go
@@ -15,26 +15,14 @@ type Config struct {
 }
 
 type Receivers struct {
-	OTLP                               config.OTLPReceiver                 `yaml:"otlp"`
-	SingletonKymaStatsReceiverCreator  *SingletonKymaStatsReceiverCreator  `yaml:"singleton_receiver_creator/kymastats,omitempty"`
-	SingletonK8sClusterReceiverCreator *SingletonK8sClusterReceiverCreator `yaml:"singleton_receiver_creator/k8s_cluster,omitempty"`
+	OTLP                              config.OTLPReceiver                `yaml:"otlp"`
+	SingletonKymaStatsReceiverCreator *SingletonKymaStatsReceiverCreator `yaml:"singleton_receiver_creator/kymastats,omitempty"`
 }
 
 type SingletonKymaStatsReceiverCreator struct {
 	AuthType                   string                     `yaml:"auth_type"`
-	LeaderElection             LeaderElection             `yaml:"leader_election"`
+	LeaderElection             metric.LeaderElection      `yaml:"leader_election"`
 	SingletonKymaStatsReceiver SingletonKymaStatsReceiver `yaml:"receiver"`
-}
-
-type SingletonK8sClusterReceiverCreator struct {
-	AuthType                    string                      `yaml:"auth_type"`
-	LeaderElection              LeaderElection              `yaml:"leader_election"`
-	SingletonK8sClusterReceiver SingletonK8sClusterReceiver `yaml:"receiver"`
-}
-
-type LeaderElection struct {
-	LeaseName      string `yaml:"lease_name"`
-	LeaseNamespace string `yaml:"lease_namespace"`
 }
 
 type SingletonKymaStatsReceiver struct {
@@ -47,32 +35,8 @@ type KymaStatsReceiver struct {
 	Resources          []ModuleGVR `yaml:"resources"`
 }
 
-type SingletonK8sClusterReceiver struct {
-	K8sClusterReceiver K8sClusterReceiver `yaml:"k8s_cluster"`
-}
-
-type K8sClusterReceiver struct {
-	AuthType               string                  `yaml:"auth_type"`
-	CollectionInterval     string                  `yaml:"collection_interval"`
-	NodeConditionsToReport []string                `yaml:"node_conditions_to_report"`
-	Metrics                K8sClusterMetricsConfig `yaml:"metrics"`
-}
-
 type MetricConfig struct {
 	Enabled bool `yaml:"enabled"`
-}
-
-type K8sClusterMetricsConfig struct {
-	// metrics allows enabling/disabling scraped metric.
-	K8sContainerStorageRequest          MetricConfig `yaml:"k8s.container.storage_request"`
-	K8sContainerStorageLimit            MetricConfig `yaml:"k8s.container.storage_limit"`
-	K8sContainerEphemeralStorageRequest MetricConfig `yaml:"k8s.container.ephemeralstorage_request"`
-	K8sContainerEphemeralStorageLimit   MetricConfig `yaml:"k8s.container.ephemeralstorage_limit"`
-	K8sContainerRestarts                MetricConfig `yaml:"k8s.container.restarts"`
-	K8sContainerReady                   MetricConfig `yaml:"k8s.container.ready"`
-	K8sNamespacePhase                   MetricConfig `yaml:"k8s.namespace.phase"`
-	K8sReplicationControllerAvailable   MetricConfig `yaml:"k8s.replication_controller.available"`
-	K8sReplicationControllerDesired     MetricConfig `yaml:"k8s.replication_controller.desired"`
 }
 
 type ModuleGVR struct {
