+++ internal/otelcollector/config/metric/gateway/receivers_test.go
@@ -84,41 +84,4 @@ func TestReceivers(t *testing.T) {
 			require.Equal(t, expectedResource.Resource, kymaStatsReceiver.Resources[i].Resource)
 		}
 	})
-
-	t.Run("singleton k8s cluster receiver creator", func(t *testing.T) {
-		gatewayNamespace := "test-namespace"
-		expectedMetricsToDrop := K8sClusterMetricsConfig{
-			K8sContainerStorageRequest:          MetricConfig{false},
-			K8sContainerStorageLimit:            MetricConfig{false},
-			K8sContainerEphemeralStorageRequest: MetricConfig{false},
-			K8sContainerEphemeralStorageLimit:   MetricConfig{false},
-			K8sContainerRestarts:                MetricConfig{false},
-			K8sContainerReady:                   MetricConfig{false},
-			K8sNamespacePhase:                   MetricConfig{false},
-			K8sReplicationControllerAvailable:   MetricConfig{false},
-			K8sReplicationControllerDesired:     MetricConfig{false},
-		}
-		collectorConfig, _, err := sut.Build(
-			ctx,
-			[]telemetryv1alpha1.MetricPipeline{
-				testutils.NewMetricPipelineBuilder().WithName("test").WithRuntimeInput(true).Build(),
-			},
-			BuildOptions{
-				GatewayNamespace: gatewayNamespace,
-			},
-		)
-		require.NoError(t, err)
-
-		singletonK8sClusterReceiverCreator := collectorConfig.Receivers.SingletonK8sClusterReceiverCreator
-		require.NotNil(t, singletonK8sClusterReceiverCreator)
-		require.Equal(t, "serviceAccount", singletonK8sClusterReceiverCreator.AuthType)
-		require.Equal(t, "telemetry-metric-gateway-k8scluster", singletonK8sClusterReceiverCreator.LeaderElection.LeaseName)
-		require.Equal(t, gatewayNamespace, singletonK8sClusterReceiverCreator.LeaderElection.LeaseNamespace)
-
-		k8sClusterReceiver := singletonK8sClusterReceiverCreator.SingletonK8sClusterReceiver.K8sClusterReceiver
-		require.Equal(t, "serviceAccount", k8sClusterReceiver.AuthType)
-		require.Equal(t, "30s", k8sClusterReceiver.CollectionInterval)
-		require.Len(t, k8sClusterReceiver.NodeConditionsToReport, 0)
-		require.Equal(t, expectedMetricsToDrop, k8sClusterReceiver.Metrics)
-	})
 }
